Reputation bug corrected
There was a bug in the reputation calculation. Trying to do the sum of the trust values, when checking the smalles address, the new values were swapped using the same variables. That meant that in a given situation we could modifiy the address of contributor1 when executing the inner loop, which led to an unwanted behavior.